pyproject.toml
[project]
name = "clinical_trial_env"
version = "1.0.0"
description = "LLM-Powered Clinical Trial Data Validator OpenEnv Environment"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "openenv-core",
    "openai>=1.0.0",
]
